requirements.txt
bitbucket-api==0.5.0
PyGithub==1.43.2
